# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Generated API",
    "description": "REST API generated by Backend Creator (bcm)",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/specialFurnitureRequests": {
      "get": {
        "tags": [
          "SpecialFurnitureRequest"
        ],
        "summary": "List all specialFurnitureRequests",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search term for @bcm.searchable fields"
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated relation names to include"
          }
        ],
        "responses": {
          "200": {
            "description": "List of specialFurnitureRequests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SpecialFurnitureRequestResponse"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "limit": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SpecialFurnitureRequest"
        ],
        "summary": "Create a new specialFurnitureRequest",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecialFurnitureRequestCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "SpecialFurnitureRequest created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialFurnitureRequestDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/specialFurnitureRequests/{id}": {
      "get": {
        "tags": [
          "SpecialFurnitureRequest"
        ],
        "summary": "Get specialFurnitureRequest by key",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "include",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SpecialFurnitureRequest details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialFurnitureRequestDataResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "tags": [
          "SpecialFurnitureRequest"
        ],
        "summary": "Update specialFurnitureRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecialFurnitureRequestUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SpecialFurnitureRequest updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialFurnitureRequestDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "SpecialFurnitureRequest"
        ],
        "summary": "Partially update specialFurnitureRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecialFurnitureRequestPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SpecialFurnitureRequest updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpecialFurnitureRequestDataResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "SpecialFurnitureRequest"
        ],
        "summary": "Delete specialFurnitureRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden — requires role: \"USER\", \"ADMIN\""
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SpecialFurnitureRequest_UserInput": {
        "type": "object",
        "properties": {
          "create": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "password": {
                "type": "string"
              },
              "phone": {
                "type": "string",
                "nullable": true
              },
              "address": {
                "type": "string",
                "nullable": true
              },
              "rate": {
                "type": "number",
                "format": "float"
              },
              "role": {
                "$ref": "#/components/schemas/Role"
              },
              "isVerified": {
                "type": "boolean"
              }
            },
            "required": [
              "name",
              "email",
              "password"
            ],
            "description": "Create a related record"
          },
          "connect": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              }
            },
            "required": [
              "id"
            ],
            "description": "Connect to an existing record by unique selector"
          }
        },
        "description": "Nested input: provide either 'create' or 'connect'",
        "anyOf": [
          {
            "required": [
              "create"
            ]
          },
          {
            "required": [
              "connect"
            ]
          }
        ]
      },
      "SpecialFurnitureRequestResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SpecialFurnitureRequestDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SpecialFurnitureRequestResponse"
          }
        }
      },
      "SpecialFurnitureRequestCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/SpecialFurnitureRequest_UserInput"
          }
        },
        "required": [
          "name",
          "phone",
          "details",
          "user"
        ]
      },
      "SpecialFurnitureRequestUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/SpecialFurnitureRequest_UserInput"
          }
        },
        "required": [
          "name",
          "phone",
          "details",
          "user"
        ]
      },
      "SpecialFurnitureRequestPatch": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RequestStatus"
          },
          "adminNote": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/SpecialFurnitureRequest_UserInput"
          }
        }
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "USER",
          "ADMIN"
        ]
      },
      "RequestStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ACCEPTED",
          "REJECTED",
          "CANCELLED",
          "AVAILABLE"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}